+++ src/main/java/com/haskov/nodes/scans/IndexScan.java
@@ -51,6 +51,12 @@ public class IndexScan implements Node {
     @Override
     public List<String> prepareTables(Long tableSize) {
         String tableName = "pg_indexscan";
+        if (SQLUtils.getTableRowCount(tableName).equals(tableSize)) {
+            V2.sql("create index if not exists pg_indexscan_idx on " + tableName + " (x)");
+            V2.sql("create index if not exists pg_indexscan_idz on " + tableName + " (z)");
+            V2.sql("vacuum freeze analyze " + tableName);
+            return new ArrayList<>(List.of(tableName));
+        }
         DropTable.dropTable(tableName);
         V2.sql("create table " + tableName + " ( x integer, y integer, z integer, w integer)");
         V2.sql("insert into " + tableName + " (x, y, z, w) select generate_series(1, ?), generate_series(1, ?), " +
